Configure submodule 'github-artifact-viewer' to track main branch
This commit updates the '.gitmodules' file to specify that the 'github-artifact-viewer' submodule, located at 'submodules/tools/github-artifact-viewer', should track the main branch. This change ensures that updates to the submodule are aligned with the latest developments on its main branch, facilitating easier synchronization and management of the submodule within our project.