Set main branch tracking for multiple submodules
This commit updates the '.gitmodules' file to specify the 'main' branch as the default tracking branch for several submodules: 'shinano', 'patio-homepage', 'interactive-math', and 'github-artifact-viewer'. These changes are made to ensure that each submodule is consistently synchronized with the latest developments on their respective 'main' branches, enhancing the overall alignment and stability of our project's submodule integrations.